GUS-7437 Move get-bytecode to tools
